backend/requirements.txt
asgiref==3.9.1
Django==4.2.23
django-filter==25.1
djangorestframework==3.16.0
isort==6.0.1
psycopg2-binary==2.9.3
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
